requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
json2html==1.3.0
webapp2==2.5.2
pandas==1.2.4